Milestone 4

See merge request xjakuboc/pv-168-project!11

+++ src/main/java/cz/muni/fi/pv168/project/storage/sql/entity/mapper/IngredientMapper.java
@@ -35,7 +35,7 @@ public final class IngredientMapper implements EntityMapper<IngredientEntity, In
                 entity.guid(),
                 entity.name(),
                 unit,
-                (int) entity.caloriesPerUnit()
+                entity.caloriesPerUnit()
         );
     }
